+++ docs/CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Changelog
 
 ### next
+  * tpm2_ecdhzgen:
+      - Added option to output the cpHash for the command TPM2_CC_ECDH_ZGen.
 
   * tpm2_ecdhkeygen:
       - Added option to output the cpHash for the command TPM2_CC_ECDH_KeyGen.
-      
+  
   * tpm2_commit:
       - Added option to output the cpHash for the command TPM2_CC_Commit.